requirements.txt
Django
django-tinymce
psycopg2-binary
Pillow
django-environ
